requirements.txt
Flask
flask-cors
numpy
opencv-python
face_recognition
Pillow
face_recognition_models